Move subscribeToOneRawSearch() one level up
