+++ src/functions/searches/explore-one-tag/explore-one-tag.ts
@@ -17,7 +17,7 @@ import {
 import { DataExplorerEntry } from '~/models/search/data-explorer-entry';
 import { toDataExplorerEntry } from '~/models/search/to-data-explorer-entry';
 import { APIContext } from '../../utils';
-import { makeSubscribeToOneRawSearch } from '../subscribe-to-one-search/subscribe-to-one-raw-search';
+import { makeSubscribeToOneRawSearch } from '../subscribe-to-one-raw-search';
 import { initiateSearch } from './initiate-search';
 
 export const makeExploreOneTag = (context: APIContext) => {
